Added files and put the skeleton files for SalaryWorker and Worker Worker Class (inherits from Person): Files: Worker.java WorkerTest.java Fields: double hourlyPayRate Methods: double calculateWeeklyPay(double hoursWorked) String displayWeeklyPay(double hoursWorked) The constructor for Worker should use super() to call the constructor for Person and then go on to set the rest of the fields. calculateWeeklyPay should return the pay total. Hours under 40 are at the HourlyRate, hours above 40 are at time and a half (1.5) displayWeeklyPay should indicate the number of hours of regular pay (40) and the total and the number of hours of overtime pay and the total as well as the total combined pay
 SalaryWorker (inherits from Worker)
Files: SalaryWorker.java SalaryWorkerTest.java
Fields:
double annualSalary
Methods:
The constructor for SalaryWorker should use super() to call the constructor for Worker
double calculateWeeklyPay(double hoursWorked)
Override calculateWeeklyPay and displayWeeklyPay calculateWeeklyPay returns the
pay total. Note that the parameter hoursWorked is not used here but is retained for
polymorphism. Take the annualSalary and divide by 52 to get the weekly pay.
displayWeeklyPay should indicate that the weekly pay is a fraction of the yearly salary.